Update test to expect null values to delete existing values
The `ReverseTopicMappingService` will currently ignore null (or empty) values when mapping to a `Topic`. That was seen as the conservative choice when introducing the original version. But it not significantly limits the support for topic references since they can be added but not removed. With #42, we're planning on introducing support for deletions, which necessitates updating the existing unit test to change its expectations. We'll follow-up with updates to the code that satisfy these conditions.