Introduce test for evaluating null `UniqueKey`
While we're introducing support for overwriting topic references via the `ReverseTopicMappingService` (#42), we don't want to support null a `IAssociationTopicBindingModel.UniqueKey`, as that could suggest an issue with data binding or the client side form, as opposed to an explicit decision to delete a topic reference. In that case, a `MappingModelValidationException` should be thrown.