Validate null `IAssociatedTopicBindingModel` or `UniqueKey`
If the `IAssociatedTopicBindingModel` or its `UniqueKey` are null, we should reject the model with a `MappingModelValidationException`, as this could be caused by a data binding error or problem with the client-side markup. We want an _explicit_ instruction to remove existing values, not an _implicit_ one. This satisfies the conditions for the two new unit tests, 9b5b57e9 and 070c1b78.

An important subtlety: In addition to elevating the consequence of a validation error from a bypass to an exception, this also removes the previous condition to ignore an _empty_ `UniqueKey`, which will be necessary to satisfy the updated test (821b4c92), and ultimately support #42.