Validate a null topic reference if the `UniqueKey` is not empty
Previously, if the topic reference returned null, the an exception was thrown. That made sense as this logic could only be reached if the `UniqueKey` was defined. Now that we allow an empty `UniqueKey` (b40aad61), however, this is ambiguous. An empty topic reference could be because the `UniqueKey` was empty, or it could be because it was not empty, but the lookup failed. We only want an error to be thrown on the latter case. Adding a condition for a non-empty `UniqueKey` satisfies this requirement.

This satisfies the final condition for deleting topic references, as evaluated in 821b4c92, and required for #42.