Allow `null` legacy topic references to be saved
If the `UniqueKey` is empty, a null topic reference will be saved to the `Topic.References` or `Topic.Attributes` collection. The former is the expected use case, and is supported given the previous update (2bd311cc). The latter continues to be supported for backward compatibility. It needs to be updated to no longer use `SetInteger()`, however, since the extension method doesn't support a `null` value, and there aren't currently any plans to make the extension methods support nullable values. As such, this needs to be updated to use `SetValue()`. Again, this is only needed for backward compatibility.