Introduced the `[MapAs()]` attribute for the `ITopicMappingService`
The `[MapAs()]` attribute allows an association on a view model to specify the view model type that the association(s) should be mapped to. This can be applied e.g. to a topic reference or a collection of relationships. This is the first step in supporting Issue #41. Subsequent updates will focus on integration and, ultimately, testing.