Expose the `[MapAs()]` attribute on the `PropertyConfiguration`
The `PropertyConfiguration` collates metadata about a property, including any configuration data from annotations, as needed by the `ITopicMappingService`. This now includes a `MapAs` property which includes the type specified by the `[MapAs()]` attribute, if set. This will allow the `ITopicMappingService` to determine if an alternate type should be mapped, as required for #41.