Introduced a private overload of `MapAsync()` that accepts `type`
The `type` allows the view model type to be specified as a parameter. This is similar to the `MapAsync<T>()` method, except that it allows the type to be specified as runtime as a parameter, instead of compile time as a generic.

Much of this logic is from the previous private `MapAsync()` overload; in fact, that overload now simply does an `ITypeLookupService.Lookup()`, validates that a type is returned, and then calls the new overload to handle the cache lookup and actual mapping.

Note: This requires that a type is looked up even if there's a cached object. Currently, this may appear to be redundant. But to fully support this, we'll need to make the cache is type aware, similar to the `IHierarchicalTopicMappingService` is today. At that point, cache lookups will need to know the type of the object.

This, along with forthcoming caching, is at the heart of the `[MapAs()]` feature (#41).