Update `MappedTopicCache` to use `Type` in `cacheKey`
With the introduction of `[MapAs()]`, it's possible for the same topic to be mapped to different types of objects. As such, relying exclusively on `Topic.Id` for the cache key is no longer sufficient. Instead, it should also factor in the `Type` of the target model. To accommodate this, the cache key for the `MappedTopicCache` is updated to use `(int, Type)`.

Assembling this, along with the `MappedTopicCacheEntry`, requires some understanding of the underlying cache, so this was additionally simplified by introducing overloads to `TryGetValue()` and `GetOrAdd()` which accept `topicId`, `type`, and/or `associations` in order to construct the cache key and `MappedTopicCacheEntry`.

Technically, this was always an issue—but the only scenario where this would occur was with the root topic being mapped (e.g., by using `MapAsync<T>()`). The chance of that being referenced within the object graph in a way that wasn't type compatible was slim, however, and thus not a priority to solve.

This includes updating the references to the `MappedTopicCache` to use these new overloads.