Establish `IsInitializing` property on `MappedTopicCacheEntry`
The `IsInitializing` property tracks that a topic is currently _in the process_ of being constructed. Historically, this hasn't been necessary since construction is expected to be a fast process. As we move toward implementing #35, however, construction potentially becomes a time-consuming operation. More importantly, while circular references are permitted with properties—due to the ability to cache the references—this isn't possible with constructors, since the object won't yet be available to cache. As such, the `IsInitializing` property allows us to track whether a circular reference has occurred.