Introduced `Preregister()` method on `MappedTopicCache`
The `Preregister()` method allows a `MappedTopicCacheEntry` to be created before the `MappedTopic` has even completed construction.

This allows a placeholder for the `MappedTopic` to be created, which in turn allows us to detect and prevent circular references during construction. (Circular references are allows when mapping properties, just not constructor parameters.)

If a circular reference is detected—i.e., `Preregister()` is called on an existing topic—then a `TopicMappingException` is thrown.