Updated `MappedTopicCache` to use private collection
As we've implemented more business logic into the `MappedTopicCache` (cb41728f, b33b60cc), the methods are now more than convenience overloads. To enforce this, the collection has been moved from a base object to a private variable. This greatly focuses the interface of the `MappedTopicCache`.

This required updating the unit tests which, previously, relied on the collection methods exposed by the underlying `ConcurrentDictionary<>` in order to avoid implicating other custom methods into each test. This is now unavoidable.