Renamed `GetOrAdd()` to `Register()`
With the `ConcurrentDictionary<>` now hidden behind the `MappedTopicCache` (60ee7d72), the public interface of the `MappedTopicCache` can adopt method names that might be a bit more intuitive and consistent with the expectations of the caller. `Register()` goes nicely with the previously introduced `Preregister()` method (cb41728f).

Unlike the `GetOrAdd()` semantics, there's no need for `Register()` to return a `MappedTopicCacheEntry`. This wasn't ever actually used in the `TopicMappingService`. But it did provide a useful shortcut for the unit tests. As a result, the unit tests have needed to be updated yet again.