Refactored `SetProperty()` around "get" semantics
Previously, the `SetProperty()` called out to a number of helper methods—such as `SetCompatibleProperty()`, `SetScalarValue()`, and SetTopicReferenceAsync()`—which not only retrieved the value from `Topic`, but _also_ set the value on the target property.

As we move toward #35, we want to update this to prefer "get" semantics, where each helper function is exclusively responsible for _retrieving_ the value, and then `SetPropertyValue()` is exclusively responsible for _setting_ the value on the property. This will allow us, eventually, to introduce an e.g. `SetParameterValue()` method which will do the same thing for the constructor parameters.

To assist with this, a lot of the common logic was moved from `SetPropertyValue()` into `GetValue()`. Some of the logic needs to remain in `SetPropertyValue()`, however; for instance, `SetPropertyValue()` shouldn't attempt to set collection property, and it needs to make sure that the target property is, in fact, settable to the source type.