Introduced new `InitializeCollection()` method
Previously, the `SetCollectionValue()` assumed that the target collection is a concrete type with an empty constructor. If it wasn't already set, it attempts to initialize it as the current type.

Ideally, this should support return types that are interfaces as well, such as `IList<>` or its derivatives. To support this, I've introduced a new `InitializeCollection()` method which provides a bit more intelligence here.

There remain limitations. This still doesn't support e.g. `IEnumerable<>`, derivatives of `IList<>`, or abstract collection types. But it will, at least, attempt to initialize `IList<>` types. It also offers a central place where we can expand this logic in the future, if we want to.

Finally, this also helps pave the way toward #35, which will always need to be able to construct new lists if they occur as constructor parameters. That will ideally need to support `IEnumerable<>` as well, but we'll revisit that later.