Introduce new `GetParameterAsync()` method
The `GetParameterAsync()` method is the counterpart to the existing `SetPropertyAsync()` method (304525d6), and implements similar logic. Since the underlying functions `SetPropertyAsync()` relies on have been refactored to a) use "get" semantics, and also b) rely on the more general `ItemConfiguration` (88e06a73), they are able to support both `SetPropertyAsync()` as well as `GetParameterAsync()`.