Prevent preregistered cache instances to be returned
The `Preregister()` method initializes a `MappedTopicCacheEntry` that doesn't have a `MappedTopic` associated with it. If a caller attempts to retrieve this, it should be treated as a miss to avoid confusion. In practice, the only purpose of `Preregister()` is to prevent circular references. By only returning `MappedTopicCacheEntry`s that were created using `Register()`, this prevents the need for callers to be aware of this logic and to check for `IsInitializing`.

To support this, the `out cacheEntry` parameter was marked as nullable. But as we don't want to do null checks on success, we've marked it as `[NotNullWhen(true)]`. That way, if a caller wraps the call to `cacheEntry` in a condition based on the successful return of `TryGetValue()`, the compiler will know the value is not null.

This required updating a couple of unit tests which _don't_ include that condition, so that they can accept potentially nullable `MappedTopicCacheEntry` instances.