Incorporate constructor mapping into `MapAsync(topic, type, …)` overload
The existing `MapAsync(topic, type, …)` overload, which was first introduced as part of the `[MapAs()]` support (#41, 61de3bf4), is the ideal place to handle constructor mapping, as it's the _only_ place where new view models are created. It will continue to first check for a cached object, but if that's not found, it will continue on to initialize a new topic view model. And if that topic view model has (public) constructor parameters, those will be mapped as part of that construction.