Introduced `ConstructedTopicViewModel` for testing constructor mapping
The `ConstructedTopicViewModel` includes each of the main types of attributes—a scalar value, a topic reference, and a relationship—in the constructor, each mapped to the names of corresponding properties. In addition, there's a fourth, optional scalar value with a default value, which is expected to be retained.

This represents a fairly complex use case for constructor mapping—likely more complicated than we'd normally advice. This ensures that the properties are properly populated, and followed—including e.g. the `[Include()]` parameter on the `topicReference` parameter.

This will be evaluated in subsequent unit tests.