Introduced test for successful constructor mapping
Using the new `ConstructedTopicViewModel` (3dcbdd62), defined a fairly complex topic graph, and confirmed that the topic was correctly mapped using the constructor parameters, while also accounting for their attributes (e.g., `[AttributeKey()]`, `[DisableMapping()]`, and `[Collection(key)]`).

This is a fairly comprehensive test. Many of these types (e.g., relationships, references) and attributes (e.g., `[AttributeKey()]`) are separately tested on properties, so we don't want to focus on the core functionality. Instead, we simply want to ensure these rules are honored when applied to parameters. For that reason, we're not breaking this up into multiple tests, as we did for properties. That makes this test more brittle than preferred for unit tests, but avoids creating a dozen tests which are largely evaluating the same underlying functionality.