Ensure compiler that this will not be null
For some reason, the compiler is insisting that these can be null—but when they're checked for nulls, it insists they will never be null. That's goofy. But, in this case, we're checking that `target?TopicReference?.TopicReference` is not null is a previous assertion, so we can just use the `!` operator to assure the compiler that it needn't worry about this.