Fixed bugs introduced during xUnit.net migration
In a couple of cases, tests now fail due to errors in how the unit test syntax was migrated. In each case, I've verified the intent against the previous version to ensure that this was a mistake in the migration, and not a case of it identifying a legitimate issue.

In one case, in `TopicRepositoryBaseTest`, there was a bug in the previous unit test which resulted in it always succeeding. In migrating this to use `Assert.DoesNotContain()`, the original error was exposed.

These now all pass correctly.