Skip `Microsoft` and `System` assemblies in `DynamicTypeLookupService`
The `DynamicTypeLookupService` is intended to search OnTopic and customer implementations of view models. It is not expected that it will be used to map to classes from `Microsoft` or `System` assemblies, of which there are many. By excluding these, we should improve performance of the initial `DynamicTypeLookupService`.

In addition, this mitigates an unexpected issue we encountered when migrating to xUnit.net, where it would attempt to load the assess Visual Studio assemblies, some of which it was unable to load. It's unclear why xUnit.net triggers this, or why it wasn't able to load those assemblies. But since we don't need to be evaluating those assemblies anyway, this is an easy fix.