Established shared infrastructure fixture
In xUnit.net, each test class is constructed for each test run. That means constructors that establish resources gets reinitialized each test, thus preventing resources from being shared between tests. That is a significant drag on performance. To mitigate that, we can establish a test fixture which allows resources to be shared between classes—and, in fact, even collections of classes. The `TopicInfrastructureFixture` does just that. It allows the caller to define what base `ITopicRepository` it wants, and then it establishes an `ITopicRepository`, `CachedTopicRepository`, and `TopicMappingService`.