Ensure state is reverted after create, update, or delete operations
Because the state of the `ITopicRepository` is now shared between multiple tests, we must take care to revert state. This isn't ideal, in that it's adding additional logic to a test which could fail. It also implies that there could be race conditions with state. As such, we may want to revisit this later. For now, it dramatically improves performance for these classes, while having minimal impact on these particular classes.

It's worth noting that other test classes, such as the `TopicRepositoryBaseTest`, require a _lot_ more cleanup and may not be worth it. For now, we're not updating that class until we're able to come up with a less invasive strategy.