Applied `TopicInfrastructureFixture` to `HiearchicalTopicMappingService` test
Originally, I had planned on introducing a new e.g. `HiearchicalTopicInfrastructureFixture` test which would introduce the `Topic` and `HiearchicalMappingService` properties to the underlying `TopicInfrastructureFixture`. This only introduces a small performance benefit, however, while introducing another class, and preventing the ability the `HierarchicalTopicMappingServiceTest` to be included in the same collection as other tests. As such, I'm instead applying the standard `TopicInfrastructureFixture` to it (a0bbb2fe), just as we did to other test classes (503bd553).