Unified tests under a new "Shared Repository" collection
The "Shared Repository" collection allows each of these tests to share the same `TopicInfrastructureFixture`, thus reducing the startup time for each class. This makes sense for these tests as they do minimal modification of state. By contrast, we do not include `ITopicRepositoryTest` in this collection because it is a bit more invasive, and we don't want to worry about potentially conflicting state.