Prefer e.g. `new Topic()` over `TopicFactory.Create()` for tests
Generally, it is preferred to use `TopicFactory.Create()`, as it will ensure that the correct strongly typed `Topic` derivative is created. For the unit tests, however, this isn't ideal. Not only does it require calling into additional code that isn't otherwise necessary, but it also requires constructing a new `DynamicTopicLookupService` for each test. The latter is reasonably fast, all things considered, but it's typically cached across an entire application; when it needs to be instantiated per test, that introduces some overhead. Since the fix is trivial, we're preferring that for unit tests.