Preferred inline initialization of properties
I had suppressed these when first migrating away from `TopicFactory.Create()` (2c031b4d) because I wanted to confirm it wouldn't interfere with the purpose of the tests, which is explicitly to establish a value and then modify it. Because the base `Topic()` constructor executes before `init` properties, this works as expected.