Applied shared fixture, collection across integration tests
Now that we're using xUnit.net, we can instantiate the `WebApplicationFactory<>` as a fixture, and share it across all integration tests using the `[Collection()]` attribute. Since the `WebApplicationFactory<>` is expensive to initialize, this dramatically reduces the execution time of the integration tests—on my machine, from ≈6.0 seconds to ≈1.5 seconds!

This also allowed me to get rid of the `IDisposable` since xUnit.net takes care of disposing the `WebApplicationFactory`.