Updated `Save()` test to use local `Topic` entities
Instead of relying on the `_topicRepository`'s existing topics, create two local topics to operate against. This prevents the need to revert state afterward—which is a bit challenging here, since the assignment of an `Id` cannot be undone. It also makes the test lighter weight and easier to reason through.