Centralize definition of `IReverseTopicMappingService`
A new instance of the `ReverseTopicMappingService` needs to be instantiated for each topic. This is a fast operation, which previously happened at the beginning of each unit test. Since xUnit.net constructs the test class for every unit test, however, we can now centralize this code as part of the class's constructor. This doesn't change the functionality, but it reduces the size and improves the readability of the `ReverseTopicMappingService` tests.