Simplified cleanup of new `Test` topic
The `Test` topic is new, so it and all of its relationships can be removed by simply deleting the topic itself. This not only is simpler, but actually more effective, as the original script failed to remove the relationships introduced by the binding model.

Since the `Test` topic is never actually added to the `_topicRepository`'s topic graph, this isn't that big of a deal, but it helps ensure that the `Test` topic doesn't linger as a reference in the `IncomingRelationships` of each of the related content types.