Introduced test for successful `Convert(object?)` to string
Whereas the previous tests evaluated converting _from_ a string _to_ a value type using `Convert(string, Type)` (e.g., 0f559142), this test evaluates the opposite direction, ensuring that a value type can successfully be converted to a string. It also evaluates a non-value type, to ensure that the default `ToString()` is called.