Centralize list validation in a new `IsList()` helper method
Currently, this is very simple, in that it exclusively checks that the `targetType` implements `IList`. This is exactly what the individual calls have been doing to date. By centralizing it, however, we allow the definition to be extended without modifying each of the pieces of code that check whether a type is a list.