Extend definition of `IsList()` to support `IList<>`, `ICollection<>`, &c.
With the bug fix for generic interfaces in place (c9bb8c65), the `InitializeList` will create a new `List<>` if the collection is null and, therefore, can support any interface that `List<>` supports—including `IList<>` (obviously), `ICollection<>`, and even `IEnumerable<>`.

This expands the interface options—and especially for mapping constructor parameters, where a) the collection type should be the simplest interface required, and b) the value will not yet be initialized. The applications for property matching are more limited, as we expect properties to be initialized by the constructor or by a default value.