Introduced unit test for initializing null collections
When using property mapping, we allow properties to be null, and to be initialized by the mapping service via their setter. This isn't a best practice, as collection properties should never be null, nor should they be settable. More importantly, this also enables constructor mapping to initialize collections, including generic interfaces, that might be used for collection parameters.

This test establishes a view model with null collections (`NullCollectionTopicViewModel`), and ensures that they can be correctly initialized and mapped. It includes a concrete generic, two generic interfaces, and a concrete collection that derives from a concrete generic.