Reintroduce `NoIndex` exclusion—but only for `Container`
In a previous commit, I fixed a bug in which `NoIndex` was treated as recursive on pages (b9433386). That's because, on a `Page`, the `NoIndex` attribute determines whether the `noindex` metadata is displayed, but doesn't affect the descendants. As such, this brought the `SitemapController` inline with the main model and expected template implementations.

One exception to that is the `Container`, which _also_ includes a `NoIndex` attribute, but which isn't inherited from `Page`. Topics of type `Container` are not expected to be rendered to web applications, and are already excluded from the `SitemapController`. Instead, `NoIndex` here applies to the children of the `Container`.

I'd still like to revisit this to determine if it makes more sense as e.g. `NoIndexTree` or `IsPrivateBranch`. But, for now, this maintains backward compatibility with the existing `NoIndex` expectations of `Container`, which were put into place to hide e.g. `Configuration` and `Administration` branches from the `SitemapController`.

For now, this satisfies the immediate bug of #38. We may want to create a new issue, however, for the proposed `NoIndexTree` functionality.