Don't overwrite initialized values with `null`
In the `TopicMappingService`, initialized values—i.e., which are assigned by a view model's constructor or auto-implemented property—are currently overwritten by `null`. I don't believe this was the behavior in previous versions; regardless, it should be treated as a bug.

_Note:_ This is different than e.g. the `Rollback()` or `Refresh()` functionality on `ITopicRepository` where we _want_ cached values in the topic graph to be overwritten; with a view model, we expect existing values will be due to initialization, not caching of potentially outdated versions. That's especially true here where we can't distinguish between an explicit `null` and an absent attribute.

This update mitigates this by explicitly checking that the new `value` is not `null` before setting the property value. This resolves #71.