Ensure that implicitly dirty attributes are `Save()`d
Previously, the `SqlTopicRepository.Save()` would save an otherwise clean `Topic` if any of its `Attributes` were marked as indexed, even though their corresponding `AttributeDescriptor` marked them as extended. That is correct behavior. The opposite scenario, however, was not being tripped. As a result, attributes reconfigured from extended to indexed storage would not be picked up as part of `Save()` unless the `Topic` was _explicitly_ marked as `IsDirty()`. This fixes that issue, and resolves #73.