Deprecate the `ListTopicViewModel`
The `List` topic is intended exclusively to support the `QueryableAttributeDescriptor` in the OnTopic Editor by allowing its `BaseTopic` to be bound to a nested topic contain, which is implemented by the OnTopic infrastructure as a `List` content type. In OnTopic 5.1.0, the need for these placeholders will be mitigated by the use of the new `[MapAs()]` attribute (#41). In anticipation of that, this placeholder view model is being marked as deprecated.