Suppress `Obsolete` warning in `TopicViewModelLookupService`
Since the `ListTopicViewModel` has been marked as deprecated (11c606f8), it will get a warning when referenced. This is a good thing. But as this is already known and accounted for, we want to disable that warning when registering the view model in the `TopicViewModelLookupService`; this will be removed once the deprecated class is removed.