Include `Version` in `TopicReferences` migration
The migration script for `TopicReferences` was written prior to versions being introduced, and thus doesn't include the version field. This introduces a problem if a topic is rolled back to a version prior to the OnTopic 5.0.0 upgrade, as that's when the `Version` for each of the `TopicReferences` will be set; in that case, the topic reference(s) will be deleted.

To mitigate that, the migration script needs to be updated to include each historical topic attribute version from `Attributes`, not just the latest version from the `AttributeIndex`, and to include the `Version` for each. This is an easy migration.

Customers that have already migrated to OnTopic 5.0.0 will need to manually migrate these and then remove duplicates. Fortunately, this only applies to a couple of migrations, and should be easy. (Other current implementations are new, and don't have this issue.)

This addresses #70.