Include `Version` in `Relationships` migration
Sets the `Version` for `Relationships` to the earliest known version of the `Source_TopicID`. This accounts for the fact that versioning is new as of OnTopic 5.0.0, and we don't have historical version data; as a result, if a `Topic` is rolled back to a version prior to the OnTopic 5.0.0 migration, we want to make sure that the historical `Relationships` aren't deleted. This addresses #74.

For this, we used `ExtendedAttributes` because the first version of a `Topic` will _always_ have an `ExtendedAttributes` record, even if it's empty. By contrast, with the migration of `Key`, `ContentType`, and `ParentID` to the `Topics` table, it isn't guaranteed that the first version of a `Topic` will have a corresponding `Attributes` record.

Customers that have already migrated to OnTopic 5.0.0 will need to manually migrate these and then remove duplicates. Fortunately, this only applies to a couple of migrations, and should be easy. (Other current implementations are new, and don't have this issue.)