Updated `Move()` escape clause to first check parent equality
The escape clause at the top of `Move()` prevents further processing of the request if the topic is being moved to the same position it is already at. There was a bug in this, however, where it failed to check if the `target` was the same as the `topic.Parent`.

That works fine so long as the `topic` and the `sibling` are, indeed, in the same collection. But if they're not, and the `topic` is in the first position, this will lead to a false positive. In effect, this prevents the first topic from being moved after a sibling in another collection.

This is easily fixed by limiting the escape clause to cases where the current parent is the same as the `target` parent. This resolves #76, though we'll want to validate this via some unit tests.