Extended `VersionHistoryIndex` to pull versions from all sources
In theory, `Attributes` should be the most authoritative place for accessing the `VersionHistory` since the OnTopic Editor will, at minimum, write a `LastModified` and `LastModifiedBy` attribute for every `Save()`. If a topic is saved programmatically, or an implementation disables the `LastModified` attributes, however, then this won't necessarily be the case.

This is especially a concern now that `Relationships` and `TopicReferences` are both versioned. We don't want updates that don't impact `Attributes` to be excluded from the `VersionHistoryIndex`—and, thus, not available for `Rollback()` in the OnTopic Editor.

This update to the `VersionHistoryIndex` addresses this by pulling versions not just from `Attributes`, but also from `ExtendedAttributes`, `Relationships`, and `TopicReferences`. This resolves #78.