Updated `StubTopicRepository` so all topics are "saved" by default
Generally, when querying an `ITopicRepository`, it's safe to assume that the topics retuned will be saved. If they aren't, there are a number of issues that can arise since `IsNew` topics will always return `IsDirty()`, among other factors. Most of the topics in the `StubTopicRepository` already have an `Id` assigned to them to simulate that they've been saved, but a number do not. This update ensures all of them have an identity assigned.

As suspected, this uncovers some issues with the unit tests. It appears at least two tests which were previously passing now fail. These have been logged as Issues #81 and #82. Those will be addressed in subsequent commits.