Fixed error in custom collection mapping test
After assigning `Id`s to all topics in the `StubTopicRepository` (45aa7bef), the unit test that validates that associations of associations don't get mapped started failing.

The reason this passed previously is because `IsNew` topics aren't cached by the `TopicMappingService`. As such, while the code does not, correctly, crawl associations of associations, it appears to do so when the topics are saved because the first association points back to the object being mapped—and, thus, when the cached instance of that mapped object is returns, its associations are mapped. This isn't because the mapping service crawled them.

This is resolved by first rearranging the assignment of the `PermittedContentTypes` (the `ContentTypes` relationship) in the `StubTopicRepository` so that both `PermittedContentTypes` themselves have `PermittedContentTypes`, and then setting it to evaluate the second of those relationships, which will not have been cached.

To help further validate this, the test now confirms that the source `Topic` is not set to the mapped `Topic`, and that it does have `PermittedContentTypes` set; this will help avoid false positives in the future.

This resolves #82.