Validate `IsNew` of `Parent` on `Save()`
The `Save()` method is obviously equipped to deal with `IsNew` topics. But it expects that the _parent_ of the current topic is saved. If not, it won't be able to save the child topic—or, if it were, it would be orphaned in the persistence store's hierarchy. Given that, I've added code to validate that the `Parent` is either `null` or `!IsNew`.

(Note: This doesn't affect recursive saves because they will save the parent before recursing through its children.)

This contributes to #77.