Updated unit test to account for `IsNew` validation in `Save()`
With the introduction of `IsNew` validation in `Save()` (4da0d431), one of the unit tests failed due to an attempt to save a topic whose parent wasn't saved. This is easily fixed by assigning `Id`s to these topics, which is what we'd expect in a real-world scenario.

While I was at it, I updated the `Assert` statement to look for an exact match, which is more precise.