Bypass `MoveTopic` if `IsNew`
While new topics can—and should—be moved in memory, they cannot—and should not—be moved in the database. This is true if the topic is new, the parent is new, or even if the sibling is new, as such operations will fail until they are saved. We choose not to throw an exception since this is a rare situation that still permits usage of the in-memory functionality of the `TopicRepository`. This satisfies the `Move()` conditions of #77.