Bypass `DeleteTopic()` if `IsNew`
While new topics can—and should—be deleted in memory, they cannot—and should not—be deleted in the database, as they don't yet exist in the database. We choose not to throw an exception since this is a rare situation that still permits usage of the in-memory functionality of the `TopicRepository`. This satisfies the `Delete()` conditions of #77.