Introduced `GetValidatedMappingType()` helper method
The private `GetValidatedMappingType()` method identifies the mapping type—either by directly passing it directly, or inferring it from the `ContentType` of a source `Topic`—and validates that it can be assigned by the `expectedType`, which is the type of the property or collection it is being assigned to. If it can, it returns the mapping type is returned; otherwise null is returned.

This will aid in validating type compatibility _before_ mapping a topic, thus reducing the overhead of using type compatibility as a shortcut for filtering, as called for in #83. This will be implemented in subsequent commits.